+++ CHANGELOG.md
@@ -28,6 +28,17 @@
 
 * Add an `anri_gender` option to control the gender of Anri of Astora.
 
+## 4.0.0-beta.4
+
+* Rename the `goal` value from "Archdragon Peak Boss" to "Archdragon Peak End
+  Boss" for consistency with other regions that have multiple bosses.
+
+* Track history in the overlay's text input.
+
+* Keep focus on the text input after sending a message through the overlay.
+
+* Unfocus and fade out the overlay when loading into a game.
+
 ## 4.0.0-beta.3
 
 * Add `Drops` and `Shops` location groups.